Added functionality to the pickups and speedmeter
You can nou actually pick the pickups. Health cannot go beyong
maxHealth, also the speed of the tank is slightly random (as it should
be) and the maximum speed now has a limit, the health pickup will give
you a random bonus as well. The speedmeter accurately shows the current
speed.